fix: int cast before === User::AUTH_PUBLIC (#956)
`User::AUTH_PUBLIC` is an int, so the config value needs to be cast to
an int before checking with `===`-operator.

This fixes a problem where having `GRAPHER_ACCESS_CUSTOMER=0` in .env is
not giving public access. Probably fixes other cases too.

+++ app/Services/Grapher/Graph/CoreBundle.php
@@ -174,7 +174,7 @@ class CoreBundle extends Graph
             return false;
         }
 
-        if( !Auth::check() && is_numeric( config( 'grapher.access.trunk' ) ) && config( 'grapher.access.trunk' ) === User::AUTH_PUBLIC ) {
+        if( !Auth::check() && is_numeric( config( 'grapher.access.trunk' ) ) && (int)config( 'grapher.access.trunk' ) === User::AUTH_PUBLIC ) {
             return $this->allow();
         }
